Renamed package to fixtures.
