--- jodd-jtx/src/test/java/jodd/jtx/data/WorkResourceManager.java
@@ -1,61 +1,0 @@
-// Copyright (c) 2003-present, Jodd Team (http://jodd.org)
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
-//
-// 1. Redistributions of source code must retain the above copyright notice,
-// this list of conditions and the following disclaimer.
-//
-// 2. Redistributions in binary form must reproduce the above copyright
-// notice, this list of conditions and the following disclaimer in the
-// documentation and/or other materials provided with the distribution.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-// POSSIBILITY OF SUCH DAMAGE.
-
-package jodd.jtx.data;
-
-import jodd.jtx.JtxResourceManager;
-import jodd.jtx.JtxTransactionMode;
-
-public class WorkResourceManager implements JtxResourceManager<WorkSession> {
-
-	int txno = 1;
-
-	public Class<WorkSession> getResourceType() {
-		return WorkSession.class;
-	}
-
-	public WorkSession beginTransaction(JtxTransactionMode jtxMode, boolean active) {
-		if (!active) {
-			return new WorkSession();
-		}
-		WorkSession work = new WorkSession(txno++);
-		work.readOnly = jtxMode.isReadOnly();
-		return work;
-	}
-
-	public void commitTransaction(WorkSession resource) {
-		resource.done();
-		txno--;
-	}
-
-	public void rollbackTransaction(WorkSession resource) {
-		resource.back();
-		txno--;
-	}
-
-	public void close() {
-	}
-
-}
